requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
azure-cognitiveservices-speech==1.44.0
azure-core==1.34.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.62.3
gtts==2.5.4
h11==0.16.0
httplib2==0.22.0
huggingface-hub==0.31.4
idna==3.10
jinja2==3.1.6
kiwisolver==1.4.8
markupsafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
proto-plus==1.26.1
protobuf==4.25.7
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.5
pydantic-core==2.33.2
pydub==0.25.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
#scipy==1.15.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
speechrecognition==3.14.3
starlette==0.46.2
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.52.0
typing-extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
ultralytics==8.3.145
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
